Change calories attribute to long type.

+++ oasp4j-samples/oasp4j-sample-core/src/main/java/io/oasp/gastronomy/restaurant/recipemanagement/dataaccess/api/RecipeEntity.java
@@ -42,7 +42,7 @@ public class RecipeEntity extends ApplicationPersistenceEntity implements Recipe
 
   private Integer cookTimeMinutes;
 
-  private Integer calories;
+  private Long calories;
 
   private String cookingInstructions;
 
@@ -188,11 +188,11 @@ public class RecipeEntity extends ApplicationPersistenceEntity implements Recipe
     this.prepTimeMinutes = prepTimeMinutes;
   }
 
-  public Integer getCalories() {
+  public Long getCalories() {
     return calories;
   }
 
-  public void setCalories(Integer calories) {
+  public void setCalories(Long calories) {
     this.calories = calories;
   }
